feat(auth): :sparkles: finish login with credentials and the relation between users and tasks was created

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup'
+import { getUserSessionServer } from '../auth/actions/auth-actions'
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url)
@@ -39,11 +40,19 @@ export async function POST(request: Request) {
   // Con hacer un request.json() ya obtenemos el body de lo que se recibe de la peticion
   // Tener en cuenta que si en el body no llega nada al tomar el body dara un error
 
+  const user = await getUserSessionServer()
+
+  if (!user) {
+    return NextResponse.json('Unauthorized', { status: 401 })
+  }
+
   try {
     const { complete, description } = await postSchema.validate(
       await request.json()
     )
-    const todo = await prisma.todo.create({ data: { complete, description } })
+    const todo = await prisma.todo.create({
+      data: { complete, description, userId: user.id },
+    })
 
     return NextResponse.json(todo)
   } catch (error) {
@@ -53,9 +62,15 @@ export async function POST(request: Request) {
 
 // Delete completed todos
 export async function DELETE() {
+  const user = await getUserSessionServer()
+
+  if (!user) {
+    return NextResponse.json('Unauthorized', { status: 401 })
+  }
+
   try {
     await prisma.todo.deleteMany({
-      where: { complete: true },
+      where: { complete: true, userId: user.id },
     })
 
     return NextResponse.json({
